Migrate remaining files and remove unnecessary things
