+++ package.json
@@ -34,7 +34,11 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.1",
+    "@types/jest": "^30.0.0",
     "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
     "typescript": "^4.9.5"
   }
 }
